refactor: refactor caseKeys
