+++ packages/mxcins-utils/src/camelcaseKeys.ts
@@ -1,38 +1,8 @@
-import { camelCase, isObject } from 'lodash';
-import mapObject from './mapObject';
+import { camelCase } from 'lodash';
+import caseKeys, { ICaseKeysOpts } from './caseKeys';
 
-const cache = new Map();
-
-export interface ICamelcaseKeysOpts {
-  /**
-   * @default false
-   *
-   */
-  readonly deep?: boolean;
-
-  /**
-   * @default []
-   */
-  readonly exclude?: Array<string | RegExp>;
-}
-
-const isObjectOrArray = (input: unknown): input is Record<string, unknown> =>
-  Array.isArray(input) || isObject(input);
-const has = (array: Array<string | RegExp>, key: string) =>
-  array.some(x => (typeof x === 'string' ? x === key : x.test(key)));
-
-const camelcaseKeys = <T = unknown>(input: unknown, opts: ICamelcaseKeysOpts = {}): T => {
-  const options = { deep: false, ...opts };
-  const { exclude } = options;
-
-  const fn = (k: string, v: unknown): [string, unknown] =>
-    exclude && has(exclude, k) ? [k, v] : [cache.get(k) || cache.set(k, camelCase(k)).get(k), v];
-
-  if (Array.isArray(input)) {
-    return (input.map(v => (isObjectOrArray(v) ? mapObject(v, fn, options) : v)) as unknown) as T;
-  }
-
-  return (isObjectOrArray(input) ? mapObject(input, fn, options) : input) as T;
+const camelcaseKeys = <T = unknown>(input: unknown, opts: ICaseKeysOpts = {}): T => {
+  return caseKeys(input, opts, camelCase);
 };
 
 export default camelcaseKeys;
